requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.1
PyPDF2==3.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1 